# OAS/movies-api.yaml
...
openapi: 3.0.0
info:
  title: Movies API
  description: This is a system API for Netflix
  contact:
    name: Cristina Akin
    email: cakin@ms3-inc.com
  version: 1.0.0
servers:
- url: https://virtserver.swaggerhub.com/cakin/movies-sys-api/1.0.0
  description: SwaggerHub API Auto Mocking
security:
- bearerAuth: []
...
paths:
  /movies:
    get:
      tags:
      - movies
      summary: get movie records from database
      description: This operation returns the movie list
      operationId: getMoviesList
      parameters:
      - name: movieName
        in: query
        description: Movie name
        required: false
        style: form
        explode: true
        schema:
          type: string
          example: Beautiful Creatures
      - name: genre
        in: query
        description: Genres used for filtering movies
        required: false
        style: form
        explode: true
        schema:
          type: string
          example: Drama
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/movie'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
    post:
      tags:
      - movies
      summary: Post a movie
      description: This operation inserts another movie record
      operationId: postMovie
      requestBody:
        $ref: '#/components/requestBodies/moviePost'
      responses:
        "201":
          description: Record created
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
  /movies/{movieId}:
    get:
      tags:
      - movies
      summary: get movie by ID
      description: This operation returns a movie record specified by ID
      operationId: getMovieByID
      parameters:
      - name: movieId
        in: path
        description: movie ID
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          example: 10
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/movie'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
    delete:
      tags:
      - movies
      summary: delete movie by ID
      description: This operation deletes a movie record specified by ID
      operationId: deleteMovieByID
      parameters:
      - name: movieId
        in: path
        description: movie ID
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          example: 10
      responses:
        "204":
          description: No content
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
    patch:
      tags:
      - movies
      summary: updates a specific movie movie record
      description: This operation updates a movie record specified by ID
      operationId: patchMovieByID
      parameters:
      - name: movieId
        in: path
        description: movie ID
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          example: 10
      requestBody:
        $ref: '#/components/requestBodies/moviePatch'
      responses:
        "204":
          description: No content
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
...
  /users:
    get:
      tags:
      - users
      summary: get all the users
      description: This operation returns the users list
      operationId: getUsersList
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/user'
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
    post:
      tags:
      - users
      summary: Create an user
      description: This operation creates another user
      operationId: postUser
      requestBody:
        $ref: '#/components/requestBodies/userPost'
      responses:
        "201":
          description: Record created
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
  /users/{userId}:
    get:
      tags:
      - users
      summary: get an user by ID
      description: This operation returns an user record specified by ID
      operationId: getUserByID
      parameters:
      - name: userId
        in: path
        description: user ID
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          example: 8
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/user'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
    delete:
      tags:
      - users
      summary: delete user by ID
      description: 'This operation deletes the user '
      operationId: deleteUserByID
      parameters:
      - name: userId
        in: path
        description: user ID
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          example: 8
      responses:
        "204":
          description: No content
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
    patch:
      tags:
      - users
      summary: updates user
      description: This operation updates an user account specified by ID
      operationId: patchUserByID
      parameters:
      - name: userId
        in: path
        description: user ID
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          example: 8
      requestBody:
        $ref: '#/components/requestBodies/userPatch'
      responses:
        "204":
          description: No content
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
...
components:
  schemas:
    movie:
      type: object
      properties:
        movieId:
          type: integer
          example: 1
        genres:
          type: array
          items:
            type: object
            allOf:
            - $ref: '#/components/schemas/genre'
        movieName:
          type: string
          example: James Bond
        releaseDate:
          type: string
          format: date
          example: 2021-10-10
        genreIds:
          type: array
          items:
            type: integer
            example: 1
      description: movie request
...
    user:
      type: object
      properties:
        userId:
          type: integer
          example: 1
        userFirstName:
          type: string
          example: Alex
        userLastName:
          type: string
          example: Smith
        email:
          type: string
          example: alex_smith@gmail.com
      description: user request
    genre:
      type: object
      properties:
        genreId:
          type: integer
          example: 1
        genreName:
          type: string
          example: Drama
      description: genre request
...
    error:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        type:
          type: string
          example: EXCEPTION
        message:
          type: string
          example: Error message in details
      description: error object
...
  requestBodies:
    moviePost:
      description: Request for post
      content:
        application/json:
          schema:
            type: object
            properties:
              genreIds:
                type: array
                items:
                  type: integer
                  example: 1
              movieName:
                type: string
                example: James Bond
              releaseDate:
                type: string
                format: date
                example: 2021-10-10
      required: true
    moviePatch:
      description: Request for patch
      content:
        application/json:
          schema:
            type: object
            properties:
              genreIds:
                type: array
                items:
                  type: integer
                  example: 1
              movieName:
                type: string
                example: James Bond
              releaseDate:
                type: string
                format: date
                example: 2021-10-10
      required: false
...
    userPost:
      description: Request for users
      content:
        application/json:
          schema:
            type: object
            properties:
              userFirstName:
                type: string
                example: Alex
              userLastName:
                type: string
                example: Smith
              email:
                type: string
                example: alex_smith@gmail.com
      required: true
    userPatch:
      description: Request for users
      content:
        application/json:
          schema:
            type: object
            properties:
              userFirstName:
                type: string
                example: Alex
              userLastName:
                type: string
                example: Smith
              email:
                type: string
                example: alex_smith@gmail.com
      required: false
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT